requirements.txt
pytest==6.2.5
pandas~=1.3.5
beautifulsoup4~=4.10.0
requests~=2.27.1
attrs~=19.3.0
jsonschema~=4.4.0
selenium==4.1.0
mysql-connector-python==8.0.22
names==0.3.0
numpy~=1.22.0
password~=0.2
phone~=0.4.3
allure-pytest==2.9.45
pytest-xdist==2.4.0
